5주차/requirements.txt
# 4주차 멀티모달 AI 실험 필수 패키지

# 기본 딥러닝 프레임워크
torch>=1.12.0
torchvision>=0.13.0

# CLIP 관련
clip-by-openai>=1.0
open_clip_torch>=2.0.0

# BLIP-2 관련
transformers>=4.20.0
accelerate>=0.20.0

# 이미지 처리
Pillow>=9.0.0
opencv-python>=4.6.0

# 데이터 처리 및 시각화
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0

# 유틸리티
requests>=2.28.0
tqdm>=4.64.0
jupyter>=1.0.0

# 선택적 패키지 (성능 향상을 위해)
# tensorboard>=2.10.0  # 실험 로깅
# wandb>=0.13.0        # 실험 추적 